Add NewCmd constructor that allows specifying the full command
Instead of hardcoding the command to use, this new constructor accepts a
slice that contains the full command. The existing New constructor now
uses this new constructor for the actual work.

Any callers that want to customize the command are now responsible for
constructing the entire command, including allocating and closing a pty
if needed.

Since allocating a pty is now optional, its cleanup in Exit() is also
made conditional.