Properly parse tokens in MI output
Tokens are used to relate MI output to commands sent. They should be
always be present, but can also occur in async records if they are the
direct result of a command.

This commit allows tokens to appear on async records as well, and allows
ommitting them on result records as well (since the documentation
indicates that the token is technically optional for result records).

Note that for async records, this token is parsed but not returned,
since there is no way for callers to know the token for commands they
send anyway. This could be added later if needed.