Fix race condition when exiting gdb
After sending the gdb-exit command, the code would call Wait() to wait
for gdb to exit. However, Wait() also frees up resources associated with
the process and its pipes, so any attempts to read from the stdout pipe
after Wait() has completed results in:

	panic: read |0: bad file descriptor

This commit adds a recordReaderDone channel, which the recordReader
writes to when it finishes. This allows the Exit() method to wait for
reading to complete before calling Wait(), preventing this race
condition.