Avoid unnecessary argument quotation
Quoting blindly every argument causes GDB to misbehave in certain situations
(bug or intended behavior?) for example `-break-delete "1"` is different from
`-break-delete 1`. It seems that GDB uses the literal raw string for some
commands (like in the `break-delete` case) but not for others; for example
`-data-evaluate-expression asd` is equivalent to `-data-evaluate-expression
"asd"`.

This should fix #4.