Add CheckedSend
This is a wrapper around Send(), that additionally checks the result
returned by gdb for an error, and transforms it into a go error if so.
This can make error handling easier (but it does not belong in Send()
itself, so Send() still allows access to the exact value returned by
gdb).